refactor: ドラッグ機能を todo-drag.js に分離
mandara.jsのファイルサイズを削減するため、Todoドラッグ＆ドロップ機能を
別モジュール(js/todo-drag.js)に分離:
- setupTodoDragAndDrop関数をtodo-drag.jsに移動
- コールバックパターンを使用して依存関係を解消
- mandara.jsは1108行から1006行に削減
